Add an option ignore SONAME differences in libraries
There are rare use cases where we do not want to compare the SONAME when
testing libraries for compatiblity or diffing libraries. This adds an
option to ignore the SONAME when doing the comparison. In these cases,
we will edit the application's DT_NEEDED to point to the other library.

This reuses the show_soname_change() function and slightly changes its
meaning to not only control if the sonames are printed but also if
they are compared. There didn't seem to be any other users of this
function and slight semantic change seemed harmless.

   * doc/manuals/abicompat.rst - added new option
   * doc/manuals/abidiff.rst - added new option to manpage
   * tools/abicompat.cc - added new command line option and passed it
   to the diff_context.
   * tools/abidiff.cc - added new command line option and passed it
   to the diff_context.
   * src/abg-comparison.cc - don't bother comparing the sonames if
   you aren't going to print them.

Signed-off-by: Ben Woodard <woodard@redhat.com>