test-read-dwarf: Use abidw rather than using the library
When this test fails, sometimes it's difficult to reproduce the
problem using abidw.  That's because this test is using the library
directly, so the exact setup is not always straightforward to
reproduce and eventually debug.

This test changes tests/test-read-dwarf.cc to make it use abidw rather
than using the library directly.

	* tests/test-read-dwarf.cc (set_suppressions)
	(set_suppressions_from_headers): Remove these static functions.
	(test_task_dwarf::perform): Use abidw rather than using the
	library.  Remove the path and architecture from the abixml files.
	Also, when the test fails, display the actual command that failed.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>