dwarf-reader: Don't assume ODR when canonicalizing DIEs
When canonicalizing DIEs, if two type DIEs have the same textual
representation and we are looking at a C++ DIE, then we assume the two
DIEs are equivalent because of the "One Definition Rule" of C++.
There are binaries however where their seem to be types DIEs that are
slightly different and yet they have the same textual representation.
It seems that this leads to some heisenbugs depending on which of the
two DIEs is "kept".  That order would then depend on whatever the
linker's output is at a particular moment when linking the binary.
Those slight differences (i.e, modulo typedefs, for instance) are a
pain to reproduce across binaries that are re-linked.  So I am
removing this optimization.  As it doesn't seem to be so useful
anymore, given all the areas of the pipeline that go improved
recently.

After doing that, it appeared the comparison engine was seeing more
DIE kinds.  So rather than aborting when seeing an unexpected DIE
kind, we now assume the two DIEs we are seeing are different.  They
are thus not canonicalized.  That's not a problem because the IR level
type canonicalizer will pick that up and fully canonicalize the type
at that level.

So as that is in place, it appeared we were doing canonical DIE
propagation even in cases where we should not.  Namely, when a subset
of the stack of aggregates being compared depends on a recursive (or
redundant) sub-type, no canonical type DIE propagation should be
performed.  The patch thus detects that state of things and disables
the canonical type DIE propagation in that case.

Because a number of part of the pipeline rely on the previously broken
state of this type DIE canonicalization process, some regression tests
are now (temporarily) broken.  This patch adapts their output,
temporarily.  Subsequent patches are going to fix the parts of the
pipeline that rely on this broken behaviour.

	* src/abg-dwarf-reader.cc (read_context::{compute_canonical_die,
	get_canonical_die, get_or_compute_canonical_die}): Don't
	special-case ODR-relevant cases.
	(is_canon_type_to_be_propagated_tag): Rename
	is_canonicalizeable_type_tag into this.
	(erase_offset_pair): Make this return a bool, not void.
	(have_offset_pair_in_common, try_canonical_die_comparison)
	(notify_die_comparison_failed): Define new static functions.
	(ABG_RETURN, ABG_RETURN_FALSE): Define new macros.
	(compare_dies): Handle DW_TAG_{class,unspecified}_type DIES.  Also
	if a comparing a DIE kind is not supported, assume the comparison
	yields "false" for the purpose of canonicalization.  Add a new
	parameter for redundant aggregates being compared.  Use the
	redundant aggregate set to detect if a redundant aggregate has
	been detected and is still being "explored"; if that's the case,
	then canonical DIE propagation is de-activated.  This might incur
	a performance hit.  Use the new ABG_RETURN and ABG_RETURN_FALSE
	macros.  Use the new try_canonical_die_comparison to compare
	canonical DIEs rather than doing it by hand.
	* tests/data/test-annotate/test14-pr18893.so.abi: Adjust.
	* tests/data/test-annotate/test15-pr18892.so.abi: Likewise.
	* tests/data/test-annotate/test17-pr19027.so.abi: Likewise.
	* tests/data/test-annotate/test18-pr19037-libvtkRenderingLIC-6.1.so.abi: Likewise.
	* tests/data/test-annotate/test19-pr19023-libtcmalloc_and_profiler.so.abi: Likewise.
	* tests/data/test-annotate/test20-pr19025-libvtkParallelCore-6.1.so.abi: Likewise.
	* tests/data/test-annotate/test21-pr19092.so.abi: Likewise.
	* tests/data/test-diff-filter/test31-pr18535-libstdc++-report-0.txt: Likewise.
	* tests/data/test-diff-filter/test31-pr18535-libstdc++-report-1.txt: Likewise.
	* tests/data/test-diff-pkg/nmap-7.70-5.el8_testjcc.x86_64-self-check-report-0.txt:
	Likewise.  This is now wrong, but I need to update it temporarily
	as a subsequent patch will right this wrong.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-0.txt:
	Likewise.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-1.txt:
	Likewise.
	* tests/data/test-read-dwarf/PR22015-libboost_iostreams.so.abi: Likewise.
	* tests/data/test-read-dwarf/PR22122-libftdc.so.abi: Likewise.
	* tests/data/test-read-dwarf/test-libaaudio.so.abi: Likewise.
	* tests/data/test-read-dwarf/test-libandroid.so.abi: Likewise.
	* tests/data/test-read-dwarf/test10-pr18818-gcc.so.abi: Likewise.
	* tests/data/test-read-dwarf/test11-pr18828.so.abi: Likewise.
	* tests/data/test-read-dwarf/test12-pr18844.so.abi: Likewise.
	* tests/data/test-read-dwarf/test14-pr18893.so.abi: Likewise.
	* tests/data/test-read-dwarf/test15-pr18892.so.abi: Likewise.
	* tests/data/test-read-dwarf/test16-pr18904.so.abi: Likewise.
	* tests/data/test-read-dwarf/test17-pr19027.so.abi: Likewise.
	* tests/data/test-read-dwarf/test18-pr19037-libvtkRenderingLIC-6.1.so.abi: Likewise.
	* tests/data/test-read-dwarf/test19-pr19023-libtcmalloc_and_profiler.so.abi: Likewise.
	* tests/data/test-read-dwarf/test20-pr19025-libvtkParallelCore-6.1.so.abi: Likewise.
	* tests/data/test-read-dwarf/test21-pr19092.so.abi: Likewise.
	* tests/data/test-read-dwarf/test22-pr19097-libstdc++.so.6.0.17.so.abi: Likewise.
	* tests/data/test-read-dwarf/test9-pr18818-clang.so.abi: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>