Add github actions to support workflows
To facilitate collaboration with developers working on the BUILD-SI
project who are using github, I have been maintaining a clone of the
libabigail repo there. Having this repo also allows us to leverage
some of the tooling that github provides, in particular the ability to
have github actions test patches to verify that the builds succeed,
and all the regression tests run successfully. This will allow it to
better integrate with the normal agile workflow used by at least this
community of developers.

While containerizing the workflow we found that fedabipkgdiff had
a subtle implicit dependency on system provided mimetypes. We found
it expedient to add logic that eliminated that dependency.

	* add .github directory
	* add .github/workflows directory
	* add .github/workflows/build-container.yaml
	* add .github/workflows/libabigail.yaml
	* add .github/workflows/test.yaml
	* add docker directory
	* add docker/Dockerfile.fedora
	* add docker/Dockerfile.ubuntu
	* add docker/Dockerfile.fedora-base
	* add docker/Dockerfile.test
	* add README-DOCKER.md
	* tools/fedabipkgdiff

Signed-off-by: Vanessa Sochat <vsochat@gmail.com>
Reviewed-by: Ben Woodard <woodard@redhat.com>