tools-utils: `entry_of_file_with_name' returns incorrect result
entry_of_file_with_name uses `string_ends_with' to test if the
filename given in the `fts_path' member of `FTSENT' matches the
`fname' argument.  This result is not correct when in the current
directory there are file with names: "./rmdir-xyx", "./dir-xyz" it
returns true for both files, this patch fixes this ambiguity by using
`basename' instead of `string_ends_with'.

	* src/abg-tools-utils.cc (entry_of_file_with_name): Replace
	call `string_ends_with' by `basename'.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>