Move find_section to exported API
While working on a project that makes use of libabigail, one of the
developers asked for the ELF find_section functions to be exported so
that they could be used. The C++ interface for these functions already
exists in libabigail and it is much easier to use than using libelf
directly. So it seems like a reasonable ask to make the existing
interface public.

This does change libabigail's ABI as well as the API. It moves the
functions from abigail::elf_helpers::find_section to
abigail::elf_readers::find_section. This of course changes the mangled
name of the functions. Therefore the library version will need to be
incremented before release.

  * src/abg-elf-helpers.h - remove the find_section functions from
    private header file.
  * include/abg-elf-reader-common.h - add find_section to public
    header file

Signed-off-by: Ben Woodard <woodard@redhat.com>