Make symtab_reader part of public API
An application is being written which makes use of libabigail and it
needs to go through the ELF symtab for many of the same reasons that
libabigail does. Using libelf directly is complicated and the
implementation of many of the features that this application needs
already exist in libabigail and so rather than duplicating code, it
seems better to just make the c++ classes that already do the job
publicly availble.

This obviously changes the public API of the library and therefore the
ABI and so the library version will need to be incremented before
release.

Signed-off-by: Ben Woodard <woodard@redhat.com>