Add github actions to support workflows
To facilitate collaboration with developers working on the BUILD-SI
project who are using github, I have been maintaining a clone of the
libabigail repo there. Having this repo also allows us to leverage
some of the tooling that github provides, in particular the ability to
have github actions test patches to verify that the builds succeed,
and all the regression tests run successfully. This will allow it to
better integrate with the normal agile workflow used by at least this
community of developers.

	* .github/workflows/build-container.yaml: New file.
	* .github/workflows/libabigail.yaml: Likewise.
	* .github/workflows/test.yaml: Likewise.
	* .github/workflows/test-fedora.yaml: Likewise.
	* .github/README.md: Likewise.
	* docker/Dockerfile.fedora: Likewise.
	* docker/Dockerfile.ubuntu: Likewise.
	* docker/Dockerfile.fedora-base: Likewise.
	* docker/Dockerfile.test: Likewise.
	* README-DOCKER.md: Likewise.

Signed-off-by: Vanessa Sochat <sochat1@llnl.gov>
Reviewed-by: Ben Woodard <woodard@redhat.com>
Signed-off-by: Dodji Seketeli <dodji@redhat.com>