abicompat: Support reading CTF and abixml
While abidw generates abidw and abidiff can consume that abidw
abicompat could only read from ELF objects with DWARF. To bring it in
line with the capabilities of abidiff, this patch adds the ability for
abicompat to read both abixml and CTF. The reason why being able to
handle abixml as well as an ELF file with DWARF is that it allows
someone to archive the abixml for an object and use it in place of
having to keep the entire object when doing abi compatibility studies.

The ability to handle CTF was also in the code that I copied over from
abidiff and so I brought it over at the same time. This should be
tested more extensively. The make check works and a trivial test using
the same object seems to work but I do not have any CTF other binaries
around to test against.

Also a feature that I asked for a long time ago was to fail if the
binaries don't have debug infomation which is needed for proper
comparison. This was added to abidiff but it didn't get added to
abicompat. Since the code was in the same area, I copied it over from
abidiff at the same time.

I do not think it makes sense to try to split these three features
apart because they are just replicating pre-existing code in a
different tool.

	* tools/abicompat.cc (options::{fail_no_debug_info, use_ctf}): New
	data members.
	(options::options): Initialize them.
	(display_usage): Add a help string for the new options
	--fail-no-debug-info and --ctf.
	(parse_command_line): Support parsing the new options
	--fail-no-debug-info and --ctf.
	(read_corpus): New static function.
	(main): Use the new read_corpus in lieu of using
	dwarf_reader::read_corpus_from_elf and/or  ctf_reader::read_corpus
	and/or xml::reader::read_corpus_from_input.
	* doc/manuals/abicompat.rst: add documentation for new options.

Signed-off-by: Ben Woodard <woodard@redhat.com>
Signed-off-by: Dodji Seketeli <dodji@redhat.com>