dwarf-reader: Avoid long comparisons when resolving C++ decl-only classes
When resolving decl-only classes, it can happen that a decl-only class
has several candidate classes definitions (in the same binary) to
resolve to.

In that case, read_context::resolve_declaration_only_classes compares
the classes definitions (of the same name) against each other.  The
problem however is that at that point, we haven't done type
canonicalization yet.  So comparing these types can have a quadratic
behaviour, i.e, take forever, especially in C++.

So, if we are looking at C++ (or any language where the ODR rules)
types, we can just do away with this comparison and assume the types
are equal, as they have the same name.

After canonicalization (which comes later), ODR violations can still
be detected (if we ware about that at all) anyway.

This patch does away with comparing aggregate types (struct/classes)
for languages that support the ODR.

In so doing, it avoids taking forever while loading the ABI corpus for
the libdyninstAPI.so.12 library from Fedora 36.

	* src/abg-dwarf-reader.cc
	(read_context::resolve_declaration_only_classes): Don't compare
	same-name-types of a binary if the ODR is relevant.
	* tests/data/test-annotate/test14-pr18893.so.abi: Adjust.
	* tests/data/test-read-dwarf/test-libandroid.so.abi: Likewise.
	* tests/data/test-read-dwarf/test14-pr18893.so.abi: Likewise.
	* tests/data/test-read-dwarf/test9-pr18818-clang.so.abi: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>