writer: Make sorting referenced typedefs types stable in abixml
After the latest commits, it apears that sorting typedefs referenced
by by other types that are emitted in the abixml file is not stable
when the underlying types are integral types.

This happens after the commit:

	commit 7cd83740bfe7e58c03ea2a66386dbb0962f7f6de
	Author: Dodji Seketeli <dodji@redhat.com>
	Date:   Fri Jul 22 22:45:04 2022 +0200

	    ir: Consider integral types of same kind and size as equivalent

This patch fixes that by using the non-internal pretty representation
(which disambiguates integral types) of types for sorting purposes.

	* src/abg-writer.cc (read_context::type_ptr_cmp::operator()): In
	the less-than operator the type pointer comparison functor, use
	the non-internal pretty representation of types for sorting
	purposes.
	* tests/data/test-annotate/test18-pr19037-libvtkRenderingLIC-6.1.so.abi:
	Adjust
	* tests/data/test-read-dwarf/test18-pr19037-libvtkRenderingLIC-6.1.so.abi:
	Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>