ctf-reader: looks for debug information in out-of-tree modules
The archive `vmlinux.ctfa' contain CTF debug information for
all the types used by more than one module, CTF for the core
kernel and CTF for each module compiled in Linux tree directory.
CTF information for out-of-tree module is not present in
`vmlinux.ctfa' file, even so, the compiler can emit the `.ctf'
section into the out-of-tree modules and it can be extracted
by the libabigail tools.

	* src/abg-ctf-reader.cc (process_ctf_archive, read_corpus
	slurp_elf_info): Avoid looking for `vmlinux.ctfa' when we aren't
	processing a `cur_corpus_group_'. So CTF info is embedded in the
	`.ko' file.
	* tests/data/Makefile.am: Add test inputs and expected files.
	* tests/data/test-read-ctf/test-linux-module.{ko,c,abi}: Add new
	test input and reference kABI.
	* tests/test-read-ctf.cc: Add new testcase.

Signed-off-by: Guillermo E. Martinez <guillermo.e.martinez@oracle.com>
Signed-off-by: Dodji Seketeli <dodji@redhat.com>