abidw: remove always true test in resolve_declaration_only_classes
The code that makes the last attempt to resolve declaration-only types
was protected by a conditional checking that the number of TUs for a
given type was more than 1. The previous branch checked for exactly 1.
However, the entire block is inside a conditional where the number of
TUs is guaranteed to be greater than 0.

Removing the conditional makes it clear that this branch handles all
remaining cases.

	* src/abg-dwarf-reader.cc
	(read_context::resolve_declaration_only_classes): Remove
	tautological conditional.

Signed-off-by: Giuliano Procida <gprocida@google.com>