dwarf-reader: Better handle the absence of a die->parent map
When analysing the ld.so binary, the DWARF reader drops the
description of the rtld_global_ro global variable on the floor.  This
is because it fails to get the decl that the rtld_global_ro variable
belongs to.  So it doesn't know where the put the description of
rtld_global_ro, in the IR.

The root cause is that the DWARF partial_unit that contains the
description of the type of the rtld_global_ro variable doesn't contain
any description of the language its types originate from.  The DWARF
reader needs to know the type, to know if it needs to construct a map
that associates each type/decl DIE to their parent DIE.  This is
useful to determine which namespace a given DIE belongs to.  But then
this is needed only for languages that support naming organising
devices like namespaces, namely, C++, Ada, Java etc.  For C, the DWARF
reader knows that by default, all decls/types belong to the global
namespace.  But then in this particular case, the partial_unit doesn't
carry any language information, so libabigail just drops the ball on
the floor, so there is no information for the rtld_global_ro variable.

The fix is thus to consider the absence of a DIE->PARENT map as
implying that all decls are defined in the global namespace.

	* src/abg-dwarf-reader.cc (get_scope_for_die): Consider that in
	the absence of the DIE->PARENT map, all decls are in the global.
	namespace.
	* tests/data/test-abidiff-exit/ld-2.28-210.so: New test.
	* tests/data/test-abidiff-exit/ld-2.28-211.so: Likewise.
	* tests/data/test-abidiff-exit/test-ld-2.28-210.so--ld-2.28-211.so.txt: New reference test output.
	* tests/test-abidiff-exit.cc (in_out_specs): Add the test above to
	the harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>