ir: Don't overdo canonical type propagation control when comparing classes
While looking at something else, I stumbled upon this problem.

When comparing a class, equals first calls the overload for
class_or_union to compare data members.  If we are in the process of
type canonicalization, the right hand side operand might be
"canonical-type-propagated", during that call to the overload.  In
other words, it can inherit the canonical type of the left-hand-side
operand.  The problem is that that canonical type propagation, if it
happens, is too early because this equals function still needs to
compare other things like virtual member functions etc.  So, the
original intend of the code was to erase the canonical type that might
have been propagated.  This is all and well.

The problem however is that the code /always/ erases the canonical
type of the right hand side operand, even if it was the result of the
propagation optimization long before it entered this equals function.
Oops.

This patch fixes that issue.

	* src/abg-ir.cc (equals): In the overload for const class_decl&,
	do not cancel the propagated canonical type if the propagation is
	not the result of invoking the equals overload for class_or_union.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>